renamed config to dalle-config and aiConfig, added types dir
