+++ src/types/ai-config.ts
@@ -1,0 +1,16 @@
+import { dalleConfigType, dalleImageSize } from "./dalle-config";
+
+export enum aiConfigTarget {
+    dalle = "dalle",
+    // chatgpt = "chatgpt"
+}
+
+export const aiConfigTypes = {
+    dalle: dalleConfigType
+}
+
+export const aiConfigValues = {
+    dalle: {
+        size: dalleImageSize
+    }
+}
