+++ src/types/dalle-config.ts
@@ -1,0 +1,9 @@
+export enum dalleConfigType {
+    size = "size"
+}
+
+export enum dalleImageSize {
+    small = "256x256",
+    medium = "512x512",
+    large = "1024x1024"
+}
